+++ crates/umasim/src/trainer/mod.rs
@@ -17,6 +17,8 @@ pub mod logging_trainer;
 pub mod mcts_trainer;
 pub mod ramen_handwritten_trainer;
 pub mod ramen_mcts_trainer;
+#[cfg(feature = "onnx")]
+pub mod ramen_nn_trainer;
 //pub mod mean_filter_collector_trainer;
 //pub mod neural_net_trainer;
 
@@ -26,6 +28,8 @@ pub use logging_trainer::LoggingTrainer;
 pub use mcts_trainer::MctsTrainer;
 pub use ramen_handwritten_trainer::RamenHandwrittenTrainer;
 pub use ramen_mcts_trainer::{RamenMctsTrainer, RamenSearchStages, RamenSelection};
+#[cfg(feature = "onnx")]
+pub use ramen_nn_trainer::RamenNnTrainer;
 //pub use mean_filter_collector_trainer::MeanFilterCollectorTrainer;
 //pub use neural_net_trainer::NeuralNetTrainer;